backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
numpy==1.24.3
gunicorn==20.1.0
boto3==1.26.165
python-dotenv==1.0.0


